Liquidation Saver + unfinished test

+++ contracts/interfaces/FeePool.sol
@@ -1,11 +1,9 @@
 pragma solidity ^0.8.20;
 
 interface FeePool {
-    function feesAvailable(
-        address account
-    ) external view returns (uint256, uint256);
+    function feesAvailable(address account) external view returns (uint256, uint256);
 
-    function totalRewardsAvailable() external view returns (uint);
+    function totalRewardsAvailable() external view returns (uint256);
 
     function claimOnBehalf(address claimingForAddress) external returns (bool);
